Relocate to singular package for sleeper client
