+++ java/clients/src/main/java/sleeper/clients/cli/SleeperClientIngest.java
@@ -13,11 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.clients.util;
+package sleeper.clients.cli;
 
-import sleeper.bulkimport.core.job.BulkImportJob;
+import com.amazonaws.services.sqs.AmazonSQS;
+
+import sleeper.ingest.core.job.IngestJob;
+import sleeper.ingest.core.job.IngestJobSerDe;
 
 @FunctionalInterface
-public interface SleeperClientImport {
-    void importFilesFromS3(BulkImportJob job);
+public interface SleeperClientIngest {
+    void sendFilesToIngest(IngestJob job);
+
+    private SleeperClientIngest ingestParquetFilesFromS3(String queueUrl, AmazonSQS sqsClient) {
+        return (job) -> {
+            sqsClient.sendMessage(queueUrl, new IngestJobSerDe().toJson((job)));
+        };
+    }
 }
